Fix quest manager initialization order
