+++ src/App.jsx
@@ -5,42 +5,8 @@ import { GameFormatter } from './game/utils/GameFormatter';
 import { GameEngine, getFarmExpansionCost, getAnimalHousingExpansionCost, BASE_FARM_PLOTS, MAX_FARM_PLOTS, BASE_ANIMAL_CAPACITY, MAX_ANIMAL_CAPACITY, FARM_EXPANSION_BATCH, ANIMAL_CAPACITY_STEP, BUILDING_UPGRADES } from './game/engine/GameEngine';
 import { SaveManager } from './game/engine/SaveManager';
 import { NotificationCenter } from './game/engine/NotificationCenter';
-
-const CROP_METADATA = Object.fromEntries(
-  Object.entries(CROPS).map(([key, crop]) => [key, { name: crop.name, emoji: crop.emoji, type: 'crop', cropKey: key }])
-);
-
-const SEED_METADATA = Object.fromEntries(
-  Object.entries(CROPS).map(([key, crop]) => [
-    `seed_${key}`,
-    { name: `${crop.name}種子`, emoji: crop.emoji, type: 'seed', cropKey: key },
-  ])
-);
-
-const PRODUCT_METADATA = Object.fromEntries(
-  Object.entries(ANIMAL_PRODUCTS).map(([key, product]) => [
-    key,
-    { name: product.name, emoji: product.emoji, type: 'product', animal: product.animal },
-  ])
-);
-
-const INVENTORY_METADATA = {
-  ...CROP_METADATA,
-  ...PRODUCT_METADATA,
-  ...SEED_METADATA,
-};
-
-const INVENTORY_ORDER = Object.keys(INVENTORY_METADATA);
-
-const createInitialInventory = () => {
-  const base = {};
-  INVENTORY_ORDER.forEach(key => {
-    base[key] = 0;
-  });
-  return base;
-};
-
-const DYNAMIC_QUEST_SLOTS = 3;
+import { createInitialInventory, INVENTORY_METADATA, INVENTORY_ORDER } from './game/state/InventoryState';
+import { QuestManager } from './game/quests/QuestManager';
 
 const FarmGame = () => {
   // 基本狀態
@@ -107,174 +73,7 @@ const FarmGame = () => {
   const [marketListSort, setMarketListSort] = useState('price');
   const [inventorySortMode, setInventorySortMode] = useState('value');
 
-  const questDefinitions = useMemo(() => {
-    const map = {};
-    NPCS.forEach(npc => {
-      (npc.quests || []).forEach(quest => {
-        if (quest?.id) {
-          map[quest.id] = { ...quest, npcName: npc.name };
-        }
-      });
-    });
-
-    Object.entries(dynamicQuests || {}).forEach(([questId, quest]) => {
-      if (questId && quest) {
-        map[questId] = quest;
-      }
-    });
-
-    return map;
-  }, [dynamicQuests]);
-
-  const getQuestTargetLabel = useCallback((quest) => {
-    if (!quest?.target) {
-      return '目標';
-    }
-
-    if (CROPS[quest.target]) {
-      const crop = CROPS[quest.target];
-      return `${crop.emoji} ${crop.name}`;
-    }
-
-    if (ANIMAL_PRODUCTS[quest.target]) {
-      const product = ANIMAL_PRODUCTS[quest.target];
-      return `${product.emoji} ${product.name}`;
-    }
-
-    if (FARM_SUPPLIES[quest.target]) {
-      const supply = FARM_SUPPLIES[quest.target];
-      return `${supply.emoji} ${supply.name}`;
-    }
-
-    return quest.target;
-  }, []);
-
-  const refreshDynamicQuests = useCallback((currentDay) => {
-    const existingDynamic = stateRef.current.dynamicQuests || {};
-    const existingLog = stateRef.current.questLog || {};
-
-    const preserved = {};
-    Object.entries(existingDynamic).forEach(([id, quest]) => {
-      const status = existingLog?.[id]?.status;
-      if (status === 'accepted' || status === 'ready') {
-        preserved[id] = quest;
-      }
-    });
-
-    const questGivers = ['商會佈告欄', '旅行商人', '鄰村里長', '合作社代表', '冒險者公會'];
-    const deliverTargets = [...Object.keys(CROPS), ...Object.keys(ANIMAL_PRODUCTS)];
-    const harvestTargets = Object.keys(CROPS);
-    const sellTargets = Object.keys(CROPS);
-
-    const randomBetween = (min, max) => Math.floor(Math.random() * (max - min + 1)) + min;
-    const pickOne = (list) => list[Math.floor(Math.random() * list.length)];
-
-    const usedIds = new Set(Object.keys(preserved));
-    const createQuestId = (slot) => {
-      let attempt = 0;
-      let id;
-      do {
-        id = `daily_${currentDay}_${slot}_${Math.random().toString(36).slice(2, 6)}`;
-        attempt += 1;
-      } while (usedIds.has(id) && attempt < 5);
-      usedIds.add(id);
-      return id;
-    };
-
-    const buildQuest = (slot) => {
-      const giver = pickOne(questGivers);
-      const questTypes = ['deliver', 'harvest', 'sell', 'pestClear'];
-      const type = pickOne(questTypes);
-      const questId = createQuestId(slot);
-
-      if (type === 'deliver' && deliverTargets.length > 0) {
-        const target = pickOne(deliverTargets);
-        const product = ANIMAL_PRODUCTS[target];
-        const crop = CROPS[target];
-        const isProduct = Boolean(product);
-        const label = isProduct
-          ? `${product.emoji} ${product.name}`
-          : `${crop.emoji} ${crop.name}`;
-        const basePrice = isProduct ? product.basePrice : crop.sellPrice;
-        const count = isProduct ? randomBetween(4, 12) : randomBetween(18, 45);
-        const reward = Math.max(150, Math.floor(basePrice * count * (isProduct ? 1.9 : 1.5)));
-        return {
-          id: questId,
-          type: 'deliver',
-          target,
-          count,
-          reward,
-          description: `${giver} 需要 ${count} 份${label}，協助供貨即可獲得酬勞。`,
-          npcName: giver,
-        };
-      }
-
-      if (type === 'harvest' && harvestTargets.length > 0) {
-        const target = pickOne(harvestTargets);
-        const crop = CROPS[target];
-        const count = randomBetween(10, 28);
-        const reward = Math.max(160, Math.floor(crop.sellPrice * count * (1.4 + Math.random() * 0.5)));
-        return {
-          id: questId,
-          type: 'harvest',
-          target,
-          count,
-          reward,
-          description: `${giver} 正準備市集，收成 ${count} 份${crop.emoji} ${crop.name} 就能領取獎金。`,
-          npcName: giver,
-        };
-      }
-
-      if (type === 'sell' && sellTargets.length > 0) {
-        const target = pickOne(sellTargets);
-        const crop = CROPS[target];
-        const count = randomBetween(15, 35);
-        const reward = Math.max(200, Math.floor(crop.sellPrice * count * (1.55 + Math.random() * 0.45)));
-        return {
-          id: questId,
-          type: 'sell',
-          target,
-          count,
-          reward,
-          description: `${giver} 想炒熱 ${crop.emoji} ${crop.name} 的行情，賣出 ${count} 份即可分紅。`,
-          npcName: giver,
-        };
-      }
-
-      const pestCount = randomBetween(3, 6);
-      const reward = randomBetween(180, 320);
-      return {
-        id: questId,
-        type: 'pestClear',
-        count: pestCount,
-        reward,
-        description: `${giver} 報告蟲害，協助處理 ${pestCount} 塊農地的害蟲。`,
-        npcName: giver,
-      };
-    };
-
-    const combined = { ...preserved };
-    let slotIndex = 0;
-    while (Object.keys(combined).length < DYNAMIC_QUEST_SLOTS) {
-      const quest = buildQuest(slotIndex);
-      combined[quest.id] = quest;
-      slotIndex += 1;
-    }
-
-    setDynamicQuests(combined);
-    setQuestLog(prev => {
-      if (!prev) {
-        return prev;
-      }
-      const next = { ...prev };
-      Object.keys(prev).forEach(id => {
-        if (id.startsWith('daily_') && !combined[id]) {
-          delete next[id];
-        }
-      });
-      return next;
-    });
-  }, [setDynamicQuests, setQuestLog, stateRef]);
+  const stateRef = useRef({});
 
   const notificationCenter = useMemo(() => new NotificationCenter(setNotifications), []);
   const addNotification = useCallback((message, options) => {
@@ -284,9 +83,27 @@ const FarmGame = () => {
     notificationCenter.dismiss(id);
   }, [notificationCenter]);
 
+  const questManager = useMemo(() => new QuestManager({
+    stateRef,
+    setQuestLog,
+    setDynamicQuests,
+    setInventory,
+    setMoney,
+    notifier: addNotification,
+  }), [stateRef, setQuestLog, setDynamicQuests, setInventory, setMoney, addNotification]);
+
+  const questDefinitions = useMemo(
+    () => questManager.buildQuestDefinitions(dynamicQuests),
+    [questManager, dynamicQuests],
+  );
+
   useEffect(() => {
-    refreshDynamicQuests(day);
-  }, [day, refreshDynamicQuests]);
+    questManager.refreshDaily(day);
+  }, [questManager, day]);
+
+  const recordQuestEvent = useCallback((event) => {
+    questManager.recordEvent(event, questDefinitions);
+  }, [questManager, questDefinitions]);
 
   const farmSize = Array.isArray(farm) ? farm.length : 0;
   const nextFarmExpansionCost = useMemo(() => (
@@ -320,8 +137,6 @@ const FarmGame = () => {
   });
   const [saveData, setSaveData] = useState('');
   const [loadData, setLoadData] = useState('');
-
-  const stateRef = useRef({});
 
   stateRef.current = {
     money,
@@ -400,81 +215,6 @@ const FarmGame = () => {
     notifier: addNotification,
   }), [addNotification]);
 
-  const recordQuestEvent = useCallback((event) => {
-    if (!event?.type) {
-      return;
-    }
-
-    setQuestLog(prev => {
-      if (!prev || Object.keys(prev).length === 0) {
-        return prev;
-      }
-
-      let mutated = false;
-      const next = { ...prev };
-
-      Object.entries(prev).forEach(([questId, entry]) => {
-        if (!entry || (entry.status !== 'accepted' && entry.status !== 'ready')) {
-          return;
-        }
-
-        const quest = questDefinitions[questId];
-        if (!quest) {
-          return;
-        }
-
-        const required = quest.count ?? 0;
-        const currentProgress = entry.progress ?? 0;
-        let updatedEntry = entry;
-
-        switch (quest.type) {
-          case 'harvest':
-            if (event.type === 'harvest' && event.crop === quest.target) {
-              const newProgress = Math.min(required, currentProgress + (event.amount ?? 1));
-              if (newProgress !== currentProgress || entry.status !== 'ready') {
-                updatedEntry = { ...entry, progress: newProgress };
-                if (required > 0 && newProgress >= required) {
-                  updatedEntry.status = 'ready';
-                }
-              }
-            }
-            break;
-          case 'sell':
-            if (event.type === 'sell' && event.crop === quest.target) {
-              const newProgress = Math.min(required, currentProgress + (event.amount ?? 1));
-              if (newProgress !== currentProgress || entry.status !== 'ready') {
-                updatedEntry = { ...entry, progress: newProgress };
-                if (required > 0 && newProgress >= required) {
-                  updatedEntry.status = 'ready';
-                }
-              }
-            }
-            break;
-          case 'pestClear':
-            if (event.type === 'pestClear') {
-              const newProgress = Math.min(required, currentProgress + (event.amount ?? 1));
-              if (newProgress !== currentProgress || entry.status !== 'ready') {
-                updatedEntry = { ...entry, progress: newProgress };
-                if (required > 0 && newProgress >= required) {
-                  updatedEntry.status = 'ready';
-                }
-              }
-            }
-            break;
-          default:
-            break;
-        }
-
-        if (updatedEntry !== entry) {
-          next[questId] = updatedEntry;
-          mutated = true;
-        }
-      });
-
-      return mutated ? next : prev;
-    });
-  }, [questDefinitions]);
-
   const gameEngine = useMemo(() => new GameEngine({
     stateRef,
     setters: {
@@ -516,146 +256,26 @@ const FarmGame = () => {
   }, [sprinklerLevel]);
 
   const acceptQuest = useCallback((questId) => {
-    const quest = questDefinitions[questId];
-    if (!quest) {
-      return;
-    }
-
-    const existing = questLog[questId];
-    if (existing?.status === 'accepted' || existing?.status === 'ready') {
-      addNotification('任務已在進行中！', { type: 'info' });
-      return;
-    }
-    if (existing?.status === 'completed') {
-      addNotification('這項任務已經完成過囉！', { type: 'info' });
-      return;
-    }
-
-    setQuestLog(prev => ({
-      ...prev,
-      [questId]: {
-        status: 'accepted',
-        progress: 0,
-        acceptedDay: day,
-      },
-    }));
-    addNotification(`📜 接下任務：「${quest.description}」`, { type: 'info' });
-  }, [questDefinitions, questLog, setQuestLog, addNotification, day]);
+    questManager.acceptQuest(questId, questDefinitions, questLog, day);
+  }, [questManager, questDefinitions, questLog, day]);
 
   const deliverQuest = useCallback((questId) => {
-    const quest = questDefinitions[questId];
-    if (!quest || quest.type !== 'deliver') {
-      return;
-    }
-
-    const entry = questLog[questId];
-    if (!entry || entry.status !== 'accepted') {
-      if (entry?.status === 'completed') {
-        addNotification('任務已經交付完成。', { type: 'info' });
-      } else {
-        addNotification('請先接受任務再進行交付。', { type: 'warning' });
-      }
-      return;
-    }
-
-    const needCount = quest.count ?? 0;
-    const available = inventory?.[quest.target] ?? 0;
-    if (available < needCount) {
-      addNotification('庫存不足，無法完成交付。', { type: 'warning' });
-      return;
-    }
-
-    setInventory(prev => ({
-      ...prev,
-      [quest.target]: Math.max(0, (prev?.[quest.target] || 0) - needCount),
-    }));
-    setMoney(prev => prev + (quest.reward ?? 0));
-    setQuestLog(prev => ({
-      ...prev,
-      [questId]: {
-        ...prev[questId],
-        status: 'completed',
-        progress: needCount,
-        completedDay: day,
-      },
-    }));
-
-    const targetLabel = getQuestTargetLabel(quest);
-    addNotification(`🎁 已交付 ${needCount} 份${targetLabel}，獲得 $${quest.reward}！`, { type: 'success' });
-  }, [questDefinitions, questLog, inventory, setInventory, setMoney, setQuestLog, getQuestTargetLabel, addNotification, day]);
+    questManager.deliverQuest(questId, questDefinitions, questLog, inventory, day);
+  }, [questManager, questDefinitions, questLog, inventory, day]);
 
   const claimQuestReward = useCallback((questId) => {
-    const quest = questDefinitions[questId];
-    if (!quest) {
-      return;
-    }
-
-    const entry = questLog[questId];
-    if (!entry) {
-      addNotification('請先接受任務！', { type: 'info' });
-      return;
-    }
-
-    if (entry.status !== 'ready') {
-      if (entry.status === 'completed') {
-        addNotification('任務獎勵已經領取過了。', { type: 'info' });
-      } else {
-        addNotification('仍未達成任務要求，持續加油！', { type: 'warning' });
-      }
-      return;
-    }
-
-    setMoney(prev => prev + (quest.reward ?? 0));
-    setQuestLog(prev => ({
-      ...prev,
-      [questId]: {
-        ...prev[questId],
-        status: 'completed',
-        progress: quest.count ?? prev[questId]?.progress ?? 0,
-        completedDay: day,
-      },
-    }));
-    addNotification(`💰 領取任務獎勵 $${quest.reward}！`, { type: 'success' });
-  }, [questDefinitions, questLog, setMoney, setQuestLog, addNotification, day]);
+    questManager.claimQuestReward(questId, questDefinitions, questLog, day);
+  }, [questManager, questDefinitions, questLog, day]);
 
   const activeQuests = useMemo(() => {
-    if (!questLog) {
-      return [];
-    }
-
-    return Object.entries(questLog)
-      .map(([questId, entry]) => {
-        const quest = questDefinitions[questId];
-        if (!quest || entry.status === 'completed') {
-          return null;
-        }
-
-        const required = quest.count ?? 0;
-        const inventoryCount = quest.type === 'deliver' ? (inventory?.[quest.target] ?? 0) : null;
-        const progress = quest.type === 'deliver'
-          ? Math.min(required, inventoryCount ?? 0)
-          : Math.min(required, entry.progress ?? 0);
-        const ready = entry.status === 'ready'
-          || (quest.type === 'deliver' && entry.status === 'accepted' && inventoryCount != null && inventoryCount >= required);
-
-        return {
-          id: questId,
-          quest,
-          entry,
-          progress,
-          required,
-          ready,
-          inventoryCount,
-        };
-      })
-      .filter(Boolean)
-      .sort((a, b) => {
-        if (a.ready === b.ready) {
-          return 0;
-        }
-        return a.ready ? -1 : 1;
-      });
-  }, [questLog, questDefinitions, inventory]);
+    const list = questManager.getActiveQuests(questDefinitions, questLog, inventory);
+    return list.sort((a, b) => {
+      if (a.ready === b.ready) {
+        return 0;
+      }
+      return a.ready ? -1 : 1;
+    });
+  }, [questManager, questDefinitions, questLog, inventory]);
 
   const availableDynamicQuests = useMemo(() => {
     if (!dynamicQuests) {
@@ -1915,7 +1535,7 @@ const FarmGame = () => {
                   {activeQuests.map(({ id, quest, entry, progress, required, ready, inventoryCount }) => {
                     const badgeText = ready ? '可完成' : entry.status === 'accepted' ? '進行中' : '等待交付';
                     const badgeClass = ready ? 'bg-green-100 text-green-700' : 'bg-gray-100 text-gray-600';
-                    const targetLabel = getQuestTargetLabel(quest);
+                    const targetLabel = questManager.getQuestTargetLabel(quest);
                     const needsText = required > 0 ? `需求：${targetLabel} x${required}` : '';
                     const progressText = required > 0
                       ? quest.type === 'deliver'
@@ -1979,7 +1599,7 @@ const FarmGame = () => {
                   <div className="space-y-2">
                     {availableDynamicQuests.map(({ id, quest }) => {
                       const required = quest.count ?? 0;
-                      const targetLabel = getQuestTargetLabel(quest);
+                      const targetLabel = questManager.getQuestTargetLabel(quest);
                       return (
                         <div key={id} className="border border-rose-100 rounded-lg p-3 bg-white/70">
                           <div className="flex items-start justify-between gap-2">
@@ -2855,7 +2475,7 @@ const FarmGame = () => {
                     statusLabel = '已完成';
                   }
 
-                  const targetLabel = getQuestTargetLabel(quest);
+                  const targetLabel = questManager.getQuestTargetLabel(quest);
 
                   return (
                     <div key={quest.id} className="border border-gray-200 rounded-lg p-3 bg-gray-50">
